Cleanup of imports.
- change of name from `stat_counter.py` to `statcounter.py` as that's the name in pyspark.
- Moved as much as I could into 'terminals'. Meaning modules which are _not_ depending on any others except for externals to pysparkling. There's still a lot to do here!
- Moved stuff out from pysparkling into pysparkling.sql because the SQL code just does not belong in the root!
- The most glaring example of this is the method `toDF` which is moved to sql.session.py and is being monkey-patched. Just like pyspark is doing it. (don't re-invent the wheel :)).
- Moved stuff to private (to pysparkling) modules (modules starting with `_`) which are not defined in pyspark. This helped to reduce the complexity a lot.
