+++ pysparkling/__init__.py
@@ -8,7 +8,7 @@ from .broadcast import Broadcast
 from .cache_manager import CacheManager, TimedCacheManager
 from .context import Context
 from .rdd import RDD
-from .sql.types import Row
+from .sql import Row
 from .statcounter import StatCounter
 from .storagelevel import StorageLevel
