+++ pysparkling/sql/__init__.py
@@ -1,11 +1,14 @@
 from ._row import Row
 from .column import Column
-from .context import SQLContext
+from .context import SQLContext  # , HiveContext, UDFRegistration
 from .dataframe import DataFrame, DataFrameNaFunctions, DataFrameStatFunctions
 from .group import GroupedData
 from .readwriter import DataFrameReader, DataFrameWriter
 from .session import SparkSession
 
+# from pyspark.sql.catalog import Catalog
+# from pyspark.sql.window import Window, WindowSpec
+# from pyspark.sql.pandas.group_ops import PandasCogroupedOps
 
 __all__ = [
     'SparkSession', 'SQLContext',  # 'HiveContext', 'UDFRegistration',
