+++ pysparkling/sql/_expressions/__init__.py
@@ -1,0 +1,5 @@
+from .expressions import BinaryOperation, Expression, NullSafeColumnOperation, UnaryExpression
+
+__all__ = [
+    'Expression', 'NullSafeColumnOperation', 'UnaryExpression', 'BinaryOperation',
+]
