+++ pysparkling/sql/_expressions/jsons.py
@@ -1,9 +1,9 @@
 import json
 
+from . import Expression
 from .._json import get_json_encoder
 from ..internal_utils.options import Options
 from ..internal_utils.readers.jsonreader import JSONReader
-from .expressions import Expression
 
 
 class StructsToJson(Expression):
