+++ pysparkling/sql/internal_utils/readers/csvreader.py
@@ -2,7 +2,7 @@ from functools import partial
 import itertools
 
 from ....fileio import TextFile
-from ...casts import get_caster
+from ..._casts import get_caster
 from ...internal_utils.options import Options
 from ...internal_utils.readers.utils import guess_schema_from_strings, resolve_partitions
 from ...schema_utils import infer_schema_from_rdd
