+++ pysparkling/sql/internal_utils/readers/utils.py
@@ -1,9 +1,7 @@
-from ..._row import row_from_keyed_values
 from ....fileio import File, TextFile
-from ...casts import get_caster
-from ...types import (
-    DecimalType, DoubleType, IntegerType, LongType, StringType, StructField, StructType, TimestampType
-)
+from ..._casts import get_caster
+from ..._row import row_from_keyed_values
+from ...types import DecimalType, DoubleType, IntegerType, LongType, StringType, StructField, StructType, TimestampType
 from ...utils import AnalysisException
