+++ pysparkling/sql/internal_utils/writers.py
@@ -4,11 +4,11 @@ import json
 import os
 import shutil
 
-from .._json import get_json_encoder
 from ...utils import portable_hash
-from ..casts import cast_to_string
-from ..expressions.aggregate.aggregations import Aggregation
-from ..expressions.mappers import StarOperator
+from .._casts import cast_to_string
+from .._expressions.aggregate.aggregations import Aggregation
+from .._expressions.mappers import StarOperator
+from .._json import get_json_encoder
 from ..functions import col
 from ..internal_utils.options import Options
 from ..internal_utils.readwrite import to_option_stored_value
