+++ pysparkling/sql/pandas/__init__.py
@@ -1,0 +1,5 @@
+"""
+This package includes the internal APIs for PySpark about interoperability
+between pandas, PySpark and PyArrow. This package should not be directly
+imported and used.
+"""
