+++ pysparkling/sql/tests/test_casts.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 import pytz
 
-from pysparkling.sql.casts import (
+from pysparkling.sql._casts import (
     cast_from_none, cast_to_array, cast_to_binary, cast_to_boolean, cast_to_byte, cast_to_date, cast_to_decimal,
     cast_to_float, cast_to_int, cast_to_long, cast_to_map, cast_to_short, cast_to_string, cast_to_struct,
     cast_to_timestamp, FloatType, identity
