+++ pysparkling/tests/test_stat_counter.py
@@ -1,7 +1,7 @@
 import pysparkling
+from pysparkling.sql._statcounter import ColumnStatHelper
 from pysparkling.sql.functions import col
 from pysparkling.sql.types import IntegerType, Row, StructField, StructType
-from pysparkling.sql._statcounter import ColumnStatHelper
 
 
 def test_mean():
